feat: Delete items from custom todolists

+++ challenge/todolist-v1/app.js
@@ -111,13 +111,24 @@ app.post('/', (req, res) => {
 
 app.post('/delete', (req, res) => {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (!err) {
-      console.log('Succesfully deleted checked item');
-      res.redirect('/');
-    }
-  });
+  if (listName === 'Today') {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (!err) {
+        console.log('Succesfully deleted checked item');
+        res.redirect('/');
+      }
+    });
+  } else {
+    List.findOne({ name: listName }, (err, foundList) => {
+      if (!err) {
+        foundList.items.pull(checkedItemId);
+        foundList.save();
+        res.redirect(`/${listName}`);
+      }
+    });
+  }
 });
 
 app.get('/about', (req, res) => {
